Update non-breaking dependencies
