+++ package.json
@@ -20,14 +20,13 @@
     "watch": "tsc -b tsconfig.build.json --watch"
   },
   "devDependencies": {
-    "@types/node": "^18.19.113",
-    "@vitest/coverage-v8": "~3.2.4",
-    "chevrotain": "^11.0.3",
-    "prettier": "^3.6.2",
+    "@types/node": "^22.19.8",
+    "@vitest/coverage-v8": "^4.0.18",
+    "prettier": "^3.8.1",
     "shx": "~0.4.0",
-    "tsx": "^4.20.4",
-    "typescript": "~5.8.3",
-    "vitest": "~3.2.4"
+    "tsx": "^4.21.0",
+    "typescript": "^5.9.3",
+    "vitest": "^4.0.18"
   },
   "pnpm": {
     "overrides": {
