+++ packages/language/package.json
@@ -36,14 +36,13 @@
     "clean": "shx rm -fr *.tsbuildinfo out"
   },
   "dependencies": {
-    "@chevrotain/regexp-to-ast": "^11.0.3",
-    "@chevrotain/types": "^11.0.3",
-    "@chevrotain/utils": "^11.0.3",
-    "chevrotain": "^11.0.3",
-    "chevrotain-allstar": "^0.3.1",
-    "lodash-es": "^4.17.21",
-    "lru-cache": "^11.2.2",
-    "minimatch": "^10.0.3",
+    "@chevrotain/regexp-to-ast": "^11.1.1",
+    "@chevrotain/types": "^11.1.1",
+    "@chevrotain/utils": "^11.1.1",
+    "chevrotain": "^11.1.1",
+    "lodash-es": "^4.17.23",
+    "lru-cache": "^11.2.5",
+    "minimatch": "^10.1.1",
     "vscode-languageserver": "~9.0.1",
     "vscode-languageserver-textdocument": "^1.0.12",
     "vscode-languageserver-types": "^3.17.5",
