+++ packages/playground/package.json
@@ -26,7 +26,7 @@
     "@codingame/monaco-vscode-configuration-service-override": "~17.2.1",
     "@codingame/monaco-vscode-outline-service-override": "~17.2.1",
     "@codingame/monaco-vscode-api": "~17.2.1",
-    "lz-string": "~1.5.0",
+    "lz-string": "^1.5.0",
     "pli-language": "workspace:*",
     "vscode-languageserver-protocol": "~3.17.5",
     "monaco-editor-wrapper": "~6.8.1",
@@ -39,7 +39,7 @@
     "vscode-ws-jsonrpc": "~3.4.0"
   },
   "devDependencies": {
-    "vite": "~6.3.5",
+    "vite": "^7.3.1",
     "@codingame/esbuild-import-meta-url-plugin": "~1.0.3"
   }
 }
