+++ packages/vscode-extension/package.json
@@ -116,15 +116,15 @@
     "dependencies": false
   },
   "dependencies": {
-    "@vscode/extension-telemetry": "^0.9.8",
-    "glob": "11.0.2",
+    "@vscode/extension-telemetry": "^1.4.0",
+    "glob": "^13.0.0",
     "pli-language": "workspace:*",
     "vscode-languageclient": "~9.0.1",
     "vscode-languageserver": "~9.0.1"
   },
   "devDependencies": {
-    "@types/vscode": "~1.67.0",
-    "@vscode/vsce": "^3.6.2",
-    "esbuild": "~0.25.5"
+    "@types/vscode": "1.67.0",
+    "@vscode/vsce": "^3.7.1",
+    "esbuild": "^0.27.2"
   }
 }
